Fix unfurl authorizations, cleanup and more logs.
Hopper reported the unfurl is not working, the issue is the auth token we use when we responde to the link_shared events with the unfurl requests, this makes sure the right user is retrieved from our database and  the right Slack token is used for the consequent unfurl request. Also this contains some cleanup of old not used fields in the attachment response and also adds some more logging to better understand future errors.